+++ examples/GAPickerViewOverview/Classes/ViewController.m
@@ -71,6 +71,24 @@
     _verticalPickerView.dataSource = self;
     _verticalPickerView.delegate = self;
     [self.view addSubview:_verticalPickerView];
+    
+    double delayInSeconds = 5.0;
+    dispatch_time_t popTime = dispatch_time(DISPATCH_TIME_NOW, (int64_t)(delayInSeconds * NSEC_PER_SEC));
+    dispatch_after(popTime, dispatch_get_main_queue(), ^(void){
+        _horizontalPickerView.selectionAlignment = GAPickerSelectionAlignmentRight;
+    });
+    
+    delayInSeconds = 10.0;
+    popTime = dispatch_time(DISPATCH_TIME_NOW, (int64_t)(delayInSeconds * NSEC_PER_SEC));
+    dispatch_after(popTime, dispatch_get_main_queue(), ^(void){
+        _horizontalPickerView.selectionAlignment = GAPickerSelectionAlignmentLeft;
+    });
+    
+    delayInSeconds = 15.0;
+    popTime = dispatch_time(DISPATCH_TIME_NOW, (int64_t)(delayInSeconds * NSEC_PER_SEC));
+    dispatch_after(popTime, dispatch_get_main_queue(), ^(void){
+        _horizontalPickerView.selectionAlignment = GAPickerSelectionAlignmentCenter;
+    });
 }
 
 - (void)viewDidUnload