+++ lib/GAPickerViewDelegate.h
@@ -32,18 +32,30 @@
 @protocol GAPickerViewDelegate <NSObject>
 @optional
 
-// returns width of column and height of row for each component.
+/*!
+ returns width of column and height of row for each component.
+ */
 - (CGFloat)pickerView:(GAPickerView *)pickerView widthForComponent:(NSInteger)component;
 - (CGFloat)pickerView:(GAPickerView *)pickerView rowHeightForComponent:(NSInteger)component;
 
-// these methods return either a plain NSString, a NSAttributedString, or a view (e.g UILabel) to display the row for the component.
-// for the view versions, we cache any hidden and thus unused views and pass them back for reuse.
-// If you return back a different object, the old one will be released. the view will be centered in the row rect
+/*!
+ these methods return either a plain NSString, a NSAttributedString, or a view (e.g UILabel) to display the row for the component.
+ for the view versions, we cache any hidden and thus unused views and pass them back for reuse.
+ If you return back a different object, the old one will be released. the view will be centered in the row rect
+ */
 - (NSString *)pickerView:(GAPickerView *)pickerView titleForColumn:(NSInteger)column forComponent:(NSInteger)component;
 
 //- (NSAttributedString *)pickerView:(UIPickerView *)pickerView attributedTitleForRow:(NSInteger)row forComponent:(NSInteger)component NS_AVAILABLE_IOS(6_0); // attributed title is favored if both methods are implemented
 //- (UIView *)pickerView:(UIPickerView *)pickerView viewForRow:(NSInteger)row forComponent:(NSInteger)component reusingView:(UIView *)view;
 
+/*!
+ Called by the picker view when the user selects a column in a component.
+ @param pickerView An object representing the picker view requesting the data.
+ @param columm A zero-indexed number identifying a column of component. Columns are numbered left-to-right.
+ @param component A zero-indexed number identifying a component of pickerView. Components are numbered top-to-bottom.
+ @discussion
+ To determine what value the user selected, the delegate uses the column index to access the value at the corresponding position in the array used to construct the component.
+ */
 - (void)pickerView:(GAPickerView *)pickerView didSelectColumn:(NSInteger)column inComponent:(NSInteger)component;
 
 @end